+++ __tests__/AmplifyMapLibreGeocoder.test.ts
@@ -61,6 +61,24 @@ describe("AmplifyGeocoderAPI", () => {
     expect(response.features).toHaveLength(0);
   });
 
+  test("forwardGeocode prioritizes bbox over proximity value", async () => {
+    const config = {
+      query: "a map query",
+      bbox: [
+        -123.31020325000009, 37.41870932473893, -121.55239075000021,
+        38.12753577367528,
+      ],
+      promixity: [-122.431297, 37.773972],
+    };
+    await AmplifyGeocoderAPI.forwardGeocode(config);
+    expect(
+      (Geo.searchByText as jest.Mock).mock.calls[0][1].biasPosition
+    ).toBeUndefined();
+    expect(
+      (Geo.searchByText as jest.Mock).mock.calls[0][1].searchAreaConstraints
+    ).toBeDefined();
+  });
+
   test("reverseGeocode returns some values in the expected format", async () => {
     const config = {
       query: "-123, 45",